- Finished implementing presets - Made relase compile with fewer warnings - Moved localized strings for video type to its own file - Removed pause button from main window and queue window - Made extension to NSXMLNode that makes getting a string from an XPath easier

+++ App/src/FilesTableView.m
@@ -242,7 +242,7 @@
 
 - (NSDragOperation)tableView:(NSTableView*)tv
                 validateDrop:(id <NSDraggingInfo>)info
-                 proposedRow:(int)row
+                 proposedRow:(NSInteger)row
        proposedDropOperation:(NSTableViewDropOperation)op
 {
     if(op==NSTableViewDropOn)
@@ -285,7 +285,7 @@
 }
 
 - (BOOL)tableView:(NSTableView *)aTableView acceptDrop:(id <NSDraggingInfo>)info
-            row:(int)row dropOperation:(NSTableViewDropOperation)operation
+            row:(NSInteger)row dropOperation:(NSTableViewDropOperation)operation
 {
     NSPasteboard* pboard = [info draggingPasteboard];
     NSArray *types = [NSArray arrayWithObjects:MZFilesTableRows,
